Refactor backup_pitr into two distinct CI tests: builtin vs Xtrabackup each of those two is further split by mysql80 and mysql57. The reasonsing for this split is that for 'builtin' we want to test with mysql-server, whereas for 'Xtrabackup' we want to test with percona-server. It much easier to do so on two distinct CI workflows than on a single one.
