+++ test/ci_workflow_gen.go
@@ -76,6 +76,7 @@ var (
 		"18",
 		"xb_backup",
 		"backup_pitr",
+		"backup_pitr_xtrabackup",
 		"21",
 		"22",
 		"mysql_server_vault",
@@ -131,7 +132,7 @@ var (
 	clustersRequiringXtraBackup = []string{
 		"xb_backup",
 		"xb_recovery",
-		"backup_pitr",
+		"backup_pitr_xtrabackup",
 	}
 	clustersRequiringMakeTools = []string{
 		"18",
@@ -169,6 +170,8 @@ func clusterMySQLVersions(clusterName string) mysqlVersions {
 	case clusterName == "schemadiff_vrepl":
 		return allMySQLVersions
 	case clusterName == "backup_pitr":
+		return allMySQLVersions
+	case clusterName == "backup_pitr_xtrabackup":
 		return allMySQLVersions
 	case clusterName == "tabletmanager_tablegc":
 		return allMySQLVersions
